[Bug #156010670] Fixed Login and Logout bugs.
Refer to the story description for this bug story.
