+++ src/components/pages/SignUpPage.js
@@ -1,20 +1,25 @@
-import React, { Component } from 'react'
-import SignUpForm from '../forms/SignUpForm'
-import NavBar from '../common/NavBar'
-import Footer from '../common/Footer'
+import React, { Component } from 'react';
+import { Redirect } from 'react-router-dom';
+import SignUpForm from '../forms/SignUpForm';
+import NavBar from '../common/NavBar';
+import Footer from '../common/Footer';
 
 export default class SignUpPage extends Component {
   render() {
+    const token = window.localStorage.getItem('token');
+    if (token) {
+      return <Redirect to="/dashboard" />;
+    }
     return (
       <div>
-        <NavBar/>
+        <NavBar />
         <div className="">
           <div className="container w-40">
-            <SignUpForm/> 
+            <SignUpForm />
           </div>
         </div>
-        <Footer/>
+        <Footer />
       </div>
-    )
+    );
   }
 }
